Moved agent methods from admin payments controller
