+++ Api/Services/Payments/Accounts/IAccountPaymentService.cs
@@ -4,7 +4,6 @@ using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Api.Models.Payments;
 using HappyTravel.Edo.Api.Models.Users;
 using HappyTravel.Edo.Data.Booking;
-using HappyTravel.Edo.Data.Management;
 using HappyTravel.EdoContracts.General;
 using HappyTravel.Money.Enums;
 using HappyTravel.Money.Models;
@@ -13,7 +12,6 @@ namespace HappyTravel.Edo.Api.Services.Payments.Accounts
 {
     public interface IAccountPaymentService
     {
-        Task<Result> ReplenishAccount(int accountId, PaymentData payment, Administrator administrator);
         Task<bool> CanPayWithAccount(AgentContext agentContext);
         Task<Result<AccountBalanceInfo>> GetAccountBalance(Currencies currency);
         Task<Result<string>> CaptureMoney(Booking booking, UserInfo user);
